Change return type form int to uint8_t

+++ common/dev/adm1272.c
@@ -29,7 +29,7 @@ LOG_MODULE_REGISTER(dev_adm1272);
 #define ADM1272_EIN_ROLLOVER_CNT_MAX 0x100
 #define ADM1272_EIN_SAMPLE_CNT_MAX 0x1000000
 #define ADM1272_EIN_ENERGY_CNT_MAX 0x8000
-int enable_adm1272_hsc(sensor_cfg *cfg, bool enable_flag)
+uint8_t enable_adm1272_hsc(sensor_cfg *cfg, bool enable_flag)
 {
 	uint8_t retry = 5;
 	int ret = -1;
